refactor: move admin components to components directory
