+++ frontend/src/components/web-settings/checkbox-item.js
