+++ frontend/src/pages/sys-admin/web-settings/web-settings.js
@@ -5,9 +5,9 @@ import { gettext, isPro, mediaUrl, logoPath, faviconPath, loginBGPath } from '..
 import Loading from '../../../components/loading';
 import toaster from '../../../components/toast';
 import MainPanelTopbar from '../main-panel-topbar';
-import Section from '../../common-admin/web-settings/section';
-import CheckboxItem from '../../common-admin/web-settings/checkbox-item';
-import FileItem from '../../common-admin/web-settings/file-item';
+import Section from '../../../components/web-settings/section';
+import CheckboxItem from '../../../components/web-settings/checkbox-item';
+import FileItem from '../../../components/web-settings/file-item';
 import InputItem from './input-item';
 
 import '../../../css/system-admin-web-settings.css';
